Revert "Issues/104 airfract age"
