+++ src/components/Tooltip.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { GroupInterface, PlanPartInterface } from "../utils/Interfaces";
 import { Constants } from "../utils/Constants";
-import {formatDuration, shortTimeSlots} from "../utils/Formating"
 
 import styles from "./Tooltip.module.scss";
 
@@ -10,16 +9,15 @@ interface Props {
   group?: GroupInterface;
 }
 
-const _formatDuration = (est: number | null | undefined) => {
-	let formatted = formatDuration(est, shortTimeSlots);
-	if(formatted && formatted.search("d"))
-		formatted = formatted + " (" + formatHours(est) + ")";
-	return formatted;
-}
-
+const timeSlots = [
+	['m', 60, 60],
+	['h', 60*60,24],
+	['d',24*60*60,7],
+	['w',7*24*60*60]
+];
 
 const formatEstimate = (est: number | null | undefined, item: PlanPartInterface) => {
-	const formattedEst = _formatDuration(est);
+	const formattedEst = formatDuration(est);
 	if(est && item?.workTime){
 		return formattedEst + " (" + formatPercent(Math.ceil(item?.workTime/est/10)) + ")";
 	}
@@ -34,6 +32,30 @@ const formatHours = (est: number | null | undefined) => {
 	const minutes = totalMinutes % 60;
 
 	return hours + "h" + minutes;
+}
+
+const formatDuration = (est: number | null | undefined) => {
+	if(!est)
+		return "-";
+	let formatted = "";
+	let addHours = true;
+	for(let i = timeSlots.length -1 ; i > -1; i --){
+		let slot = timeSlots[i];
+		//@ts-ignore
+		let val = Math.floor(est/slot[1]);
+		if(val == 0) {
+			addHours = i > 2 ;
+			continue;
+		}
+
+		if(slot.length > 2)
+			//@ts-ignore
+			val = val%slot[2];
+		formatted = formatted + val + slot[0];
+	}
+	if(addHours)
+		formatted = formatted + " (" + formatHours(est) + ")";
+	return formatted;
 }
 
 const formatPercent = (est: number | null | undefined) => {
@@ -56,8 +78,8 @@ const commonPlanDescription = ( item ) =>{
 			<table className={styles.performance}>
 				<tr><td style={{paddingRight:"10px"}}>Estimated work-time  </td><td>{formatEstimate(item?.estMin*3600, item)}</td></tr>
 				<tr><td>Average work-time  </td><td>{formatEstimate(scale(item?.plannedWorkTime, 1000), item)}</td></tr>
-				<tr><td>Work-time  </td><td>{_formatDuration(scale(item?.workTime, 1000))}</td></tr>
-				<tr><td>Duration  </td><td>{_formatDuration(scale(item?.duration, 1000))}</td></tr>
+				<tr><td>Work-time  </td><td>{formatDuration(scale(item?.workTime, 1000))}</td></tr>
+				<tr><td>Duration  </td><td>{formatDuration(scale(item?.duration, 1000))}</td></tr>
 				<tr><td>Mechanic count  </td><td>{item?.numberOfMechanics}</td></tr>
 			</table>
 		</div>
